split nodes.py file into categories, remove unused functions, change function names to nouns, rename hamilton_run.py to run.py
