+++ naturf/nodes.py
@@ -1,134 +1,11 @@
-import os
 import geopandas as gpd
 import math
 import numpy as np
 import pandas as pd
 from pyproj.crs import CRS
 from hamilton.function_modifiers import extract_columns
-import xarray as xr
-import struct
-
-from functools import partial
-from rasterio.enums import MergeAlg
-from geocube.api.core import make_geocube
-from geocube.rasterize import rasterize_image
 
 from .config import Settings
-
-
-def _average_north_south_building_distances(building_id: pd.Series) -> dict:
-    """Create a dictionary with an initialization value of 0 present for each building id.
-
-    :param building_id:                         Building ID field.
-    :type building_id:                          pd.Series
-
-    :return:                                    Dictionary of with an initialization value of 0 present for each
-                                                building id.
-
-    """
-
-    return {i: 0 for i in building_id.unique()}
-
-
-def _building_height_dict(building_id: pd.Series, building_height: pd.Series) -> dict:
-    """Calculate building height per building.
-
-    :param building_id:                         Building ID field.
-    :type building_id:                          pd.Series
-
-    :param building_height:                     Building height field.
-    :type building_height:                      pd.Series
-
-    :return:                                    Dictionary of building id to building height
-
-    """
-
-    df = pd.DataFrame({Settings.id_field: building_id, Settings.height_field: building_height})
-
-    return df.groupby(Settings.id_field)[Settings.height_field].max().to_dict()
-
-
-def _footprint_building_areas(_building_height_dict: dict, _plan_area_dict: dict) -> dict:
-    """Create a dictionary of the height and plan_area to a dictionary where the key is the
-    target building ID.
-
-    :param _building_height_dict:               Dictionary of building id to building height
-    :type _building_height_dict:                dict
-
-    :param _plan_area_dict:                     Dictionary of building id to the neighboring building areas.
-    :type _plan_area_dict:                      dict
-
-    :return:                                    Dictionary of building id to the height and plan area.
-
-    """
-
-    return {i: [_building_height_dict[i], _plan_area_dict[i]] for i in _plan_area_dict.keys()}
-
-
-def _plan_area_dict(building_id: pd.Series, building_area_neighbor: pd.Series) -> dict:
-    """Calculate the total area of all buildings within the plan area.
-
-    :param building_id:                         Building ID field.
-    :type building_id:                          pd.Series
-
-    :param building_area_neighbor:              Building area neighbor field.
-    :type building_area_neighbor:               pd.Series
-
-    :return:                                    Dictionary of building id to the neighboring building areas
-
-    """
-
-    df = pd.DataFrame(
-        {Settings.id_field: building_id, Settings.neighbor_area_field: building_area_neighbor}
-    )
-
-    return df.groupby(Settings.id_field)[Settings.neighbor_area_field].sum().to_dict()
-
-
-def aggregate_rasters(rasterize_parameters: xr.Dataset) -> xr.Dataset:
-    """Divide each raster by the number of buildings in the cell to get the average parameter value for each cell.
-
-    :param rasterize_parameters:                    Xr.Dataset with rasterized parameters summed at the defined resolution.
-    :type rasterize_parameters:                     Xr.Dataset
-
-    :return:                                        Xr.Dataset containing rasterized parameter values averaged at the defined resolution.
-    """
-
-    return (rasterize_parameters / rasterize_parameters["building_count"]).fillna(0)
-
-
-def angle_in_degrees_to_neighbor(
-    building_centroid_target: gpd.GeoSeries, building_centroid_neighbor: gpd.GeoSeries
-) -> pd.Series:
-    """Calculate the angle in degrees of the neighbor building orientation to the target.
-    Adjust the angle to correspond to a circle where 0/360 degrees is directly east, and the
-    degrees increase counter-clockwise.
-
-    :param building_centroid_target:            Centroid geometry from the target building geometry.
-    :type building_centroid_target:             gpd.GeoSeries
-
-    :param building_centroid_neighbor:          Centroid geometry from the neighbor building geometry.
-    :type building_centroid_neighbor:           gpd.GeoSeries
-
-    :return:                                    Angle in degrees for each building interaction.
-
-    """
-
-    # unpack coordinates for readability
-    target_x = building_centroid_target.x
-    target_y = building_centroid_target.y
-    neighbor_x = building_centroid_neighbor.x
-    neighbor_y = building_centroid_neighbor.y
-
-    # calculate the angle in degrees of the neighbor building orientation to the target
-    angle_series = np.degrees(np.arctan2(neighbor_y - target_y, neighbor_x - target_x))
-
-    # adjust the angle to correspond to a circle where 0/360 degrees is directly east, and the degrees increase
-    # counter-clockwise
-    return pd.Series(
-        np.where(angle_series < 0, angle_series + Settings.DEGREES_IN_CIRCLE, angle_series),
-        index=building_centroid_target.index,
-    )
 
 
 def area_weighted_mean_of_building_heights(
@@ -160,67 +37,6 @@ def area_weighted_mean_of_building_heights(
     df = volume_sum / area_sum
 
     return pd.Series(df.values)
-
-
-def average_building_heights(
-    building_id: pd.Series, building_height_neighbor: pd.Series
-) -> pd.Series:
-    """Average building heights for each target building when considering themselves
-    and those that are within their buffered area.
-
-    :param building_id:                         Building ID field.
-    :type building_id:                          pd.Series
-
-    :param building_height_neighbor:            Building height field for neighbors
-    :type building_height_neighbor:             pd.Series
-
-    :return:                                    Series of building heights for each target building in the buffered area
-
-    """
-
-    df = pd.DataFrame(
-        {Settings.id_field: building_id, Settings.neighbor_height_field: building_height_neighbor}
-    )
-
-    return df.groupby(Settings.id_field)[Settings.neighbor_height_field].mean().fillna(0)
-
-
-def average_direction_distance(
-    building_id: pd.Series,
-    orientation_to_neighbor: pd.Series,
-    distance_to_neighbor_by_centroid: pd.Series,
-) -> pd.Series:
-    """Calculate the average directional distances for each target building to its neighbors
-
-    :param building_id:                         Building ID field.
-    :type building_id:                          pd.Series
-
-    :param orientation_to_neighbor:             Either the east-west or north-south orientation of the target building
-                                                to its neighbors.
-    :type orientation_to_neighbor:              pd.Series
-
-    :param distance_to_neighbor_by_centroid:    Distance to neighbor from the target building using centroids.
-    :type distance_to_neighbor_by_centroid:     pd.Series
-
-    :return:                                    Series of average directional distances for each target building to its
-                                                neighbors.
-
-    """
-
-    df = pd.DataFrame(
-        {
-            Settings.id_field: building_id,
-            "orientation": orientation_to_neighbor,
-            "distance": distance_to_neighbor_by_centroid,
-        }
-    )
-
-    df = df.groupby([Settings.id_field, "orientation"])["distance"].mean().reset_index()
-
-    # exclude the target building to target building distance and direction (would be == 0)
-    df = df.loc[df["distance"] > 0]
-
-    return df
 
 
 def average_distance_between_buildings(distance_between_buildings: pd.Series) -> pd.Series:
@@ -260,56 +76,6 @@ def building_area(building_geometry: pd.Series) -> pd.Series:
     """
 
     return building_geometry.area
-
-
-def building_centroid(building_geometry: pd.Series) -> pd.Series:
-    """Calculate the centroid of the building geometry.
-
-    :param building_geometry:                       Geometry of the building.
-    :type building_geometry:                        pd.Series
-
-    :return:                                        pd.Series
-
-    """
-
-    return building_geometry.centroid
-
-
-def building_centroid_neighbor(
-    building_geometry_neighbor: pd.Series, target_crs: CRS
-) -> gpd.GeoSeries:
-    """Calculate the centroid geometry from the neighbor building geometry.
-
-    :param building_geometry_neighbor:          Geometry field for the neighboring buildings from the spatially
-                                                joined data.
-    :type building_geometry_neighbor:           pd.Series
-
-    :param target_crs:                          Coordinate reference system field of the parent geometry.
-    :type target_crs:                           pd.Series
-
-    :return:                                    The centroid geometry from the neighbor building geometry as a
-                                                GeoSeries.
-
-    """
-
-    return gpd.GeoSeries(building_geometry_neighbor, crs=target_crs).centroid
-
-
-def building_centroid_target(building_geometry_target: pd.Series, target_crs: CRS) -> gpd.GeoSeries:
-    """Calculate the centroid geometry from the parent building geometry.
-
-    :param building_geometry_target:            Geometry field for the target buildings from the spatially
-                                                joined data.
-    :type building_geometry_target:             pd.Series
-
-    :param target_crs:                          Coordinate reference system field of the parent geometry.
-    :type target_crs:                           pd.Series
-
-    :return:                                    The centroid geometry from the target building geometry as a GeoSeries.
-
-    """
-
-    return gpd.GeoSeries(building_geometry_target, crs=target_crs).centroid
 
 
 def buildings_intersecting_plan_area(
@@ -543,24 +309,6 @@ def distance_between_buildings(buildings_intersecting_plan_area: gpd.GeoDataFram
     return buildings_intersecting_plan_area.distance(neighbor_geometry).rename(
         Settings.distance_between_buildings
     )
-
-
-def distance_to_neighbor_by_centroid(
-    building_centroid_target: gpd.GeoSeries, building_centroid_neighbor: gpd.GeoSeries
-) -> pd.Series:
-    """Calculate the distance from the target building neighbor to each neighbor building centroid.
-
-    :param building_centroid_target:            Centroid geometry from the target building geometry.
-    :type building_centroid_target:             gpd.GeoSeries
-
-    :param building_centroid_neighbor:          Centroid geometry from the neighbor building geometry.
-    :type building_centroid_neighbor:           gpd.GeoSeries
-
-    :return:                                    Distance field in CRS units.
-
-    """
-
-    return building_centroid_target.distance(building_centroid_neighbor)
 
 
 @extract_columns(*[Settings.id_field, Settings.height_field, Settings.geometry_field])
@@ -834,93 +582,6 @@ def frontal_length(
     )
 
 
-@extract_columns(*Settings.spatial_join_list)
-def get_neighboring_buildings_df(
-    building_id: pd.Series,
-    building_height: pd.Series,
-    building_geometry: pd.Series,
-    building_area: pd.Series,
-    building_centroid: pd.Series,
-    total_plan_area_geometry: pd.Series,
-    target_crs: CRS,
-    join_type: str = "left",
-    join_predicate: str = "intersects",
-    join_lsuffix: str = "target",
-    join_rsuffix: str = Settings.neighbor,
-) -> gpd.GeoDataFrame:
-    """Conduct a spatial join to get the building centroids that intersect the buffered target buildings.
-
-    :param building_id:                         Building ID field.
-    :type building_id:                          pd.Series
-
-    :param building_height:                     Building height field.
-    :type building_height:                      pd.Series
-
-    :param building_geometry:                   Geometry field for the buildings.
-    :type building_geometry:                    pd.Series
-
-    :param building_area:                       Building area field.
-    :type building_area:                        pd.Series
-
-    :param building_centroid:                   Point centroid geometry of the building.
-    :type building_centroid:                    pd.Series
-
-    :param total_plan_area_geometry:            Geometry of the buffered building.
-    :type total_plan_area_geometry:             pd.Series
-
-    :param target_crs:                          Coordinate reference system field of the parent geometry.
-    :type target_crs:                           pd.Series
-
-    :param join_type:                           Type of join desired.
-                                                DEFAULT: `left`
-    :type join_type:                            str
-
-    :param join_predicate:                      Selected topology of join.
-                                                DEFAULT: `intersects`
-    :type join_predicate:                       str
-
-    :param join_lsuffix:                        Suffix of the left object in the join.
-                                                DEFAULT: `target`
-    :type join_lsuffix:                         str
-
-    :param join_rsuffix:                        Suffix of the right object in the join.
-                                                DEFAULT: `neighbor`
-    :type join_rsuffix:                         str
-
-    :return:                                    GeoDataFrame of building centroids that intersect the buffered target
-                                                buildings and their attributes.
-
-    """
-
-    # assemble a pandas data frame
-    df = pd.DataFrame(
-        {
-            Settings.id_field: building_id,
-            Settings.height_field: building_height,
-            Settings.area_field: building_area,
-            Settings.geometry_field: building_geometry,
-            Settings.centroid_field: building_centroid,
-            Settings.buffered_field: total_plan_area_geometry,
-        }
-    )
-
-    # create left and right geodataframes
-    left_gdf = gpd.GeoDataFrame(df, geometry=Settings.buffered_field, crs=target_crs)
-    right_gdf = gpd.GeoDataFrame(df, geometry=Settings.centroid_field, crs=target_crs)
-
-    # spatially join the building centroids to the target buffered areas
-    xdf = gpd.sjoin(
-        left_df=left_gdf,
-        right_df=right_gdf,
-        how=join_type,
-        predicate=join_predicate,
-        lsuffix=join_lsuffix,
-        rsuffix=join_rsuffix,
-    )
-
-    return xdf
-
-
 def grimmond_oke_displacement_height(building_height: pd.Series) -> pd.Series:
     """Calculate the Grimmond & Oke displacement height for each building
 
@@ -1105,212 +766,6 @@ def mean_building_height(buildings_intersecting_plan_area: gpd.GeoDataFrame) -> 
     return pd.Series(df.values)
 
 
-def merge_parameters(
-    frontal_area_density: pd.DataFrame,
-    plan_area_density: pd.DataFrame,
-    rooftop_area_density: pd.DataFrame,
-    plan_area_fraction: pd.Series,
-    mean_building_height: pd.Series,
-    standard_deviation_of_building_heights: pd.Series,
-    area_weighted_mean_of_building_heights: pd.Series,
-    building_surface_area_to_plan_area_ratio: pd.Series,
-    frontal_area_index: pd.DataFrame,
-    complete_aspect_ratio: pd.Series,
-    height_to_width_ratio: pd.Series,
-    sky_view_factor: pd.Series,
-    grimmond_oke_roughness_length: pd.Series,
-    grimmond_oke_displacement_height: pd.Series,
-    raupach_roughness_length: pd.DataFrame,
-    raupach_displacement_height: pd.DataFrame,
-    macdonald_roughness_length: pd.DataFrame,
-    macdonald_displacement_height: pd.Series,
-    vertical_distribution_of_building_heights: pd.DataFrame,
-    building_geometry: pd.Series,
-    target_crs: CRS,
-) -> gpd.GeoDataFrame:
-    """Merge all parameters into one Pandas DataFrame.
-
-    :param frontal_area_density:                       Frontal area density at each specified height increment and each cardinal direction.
-    :type frontal_area_density:                        pd.DataFrame
-
-    :param plan_area_density:                          Plan area density at each specified height increment.
-    :type plan_area_density:                           pd.DataFrame
-
-    :param rooftop_area_density:                       Rooftop area density at each specified height increment.
-    :type rooftop_area_density:                        pd.DataFrame
-
-    :param plan_area_fraction:                         Plan area fraction for each building.
-    :type plan_area_fraction:                          pd.Series
-
-    :param mean_building_height:                      Series of mean building height for all buildings within the target building's total plan area.
-    :type mean_building_height:                       pd.Series
-
-    :param standard_deviation_of_building_heights:     Series of standard deviation of building height for all buildings within the target building's
-                                                       total plan area.
-    :type standard_deviation_of_building_heights:      pd.Series
-
-    :param area_weighted_mean_of_building_heights:     Series of area weighted mean building heights for each building.
-    :type area_weighted_mean_of_building_heights:      pd.Series
-
-    :param building_surface_area_to_plan_area_ratio:   Series of building surface area to plan area ratio for each building.
-    :type building_surface_area_to_plan_area_ratio:    pd.Series
-
-    :param frontal_area_index:                         Frontal area index for each building in each cardinal direction.
-    :type frontal_area_index:                          pd.DataFrame
-
-    :param complete_aspect_ratio:                      Complete aspect ratio for each building.
-    :type complete_aspect_ratio:                       pd.Series
-
-    :param height_to_width_ratio:                      Height-to-width ratio for each building.
-    :type height_to_width_ratio:                       pd.Series
-
-    :param sky_view_factor:                            Sky view factor for each building.
-    :type sky_view_factor:                             pd.Series
-
-    :param grimmond_oke_roughness_length:              Grimmond & Oke roughness length for each building.
-    :type grimmond_oke_roughness_length:               pd.Series
-
-    :param grimmond_oke_displacement_height:           Grimmond & Oke displacement height for each building.
-    :type grimmond_oke_displacement_height:            pd.Series
-
-    :param raupach_roughness_length:                   Raupach roughness length for each building in each cardinal direction.
-    :type raupach_roughness_length:                    pd.DataFrame
-
-    :param raupach_displacment_height:                 Raupach displacment height for each building in each cardinal direction.
-    :type raupach_displacment_height:                  pd.DataFrame
-
-    :param macdonald_roughness_length:                 Macdonald roughness_length for each building in each cardinal direction.
-    :type macdonald_roughness_length:                  pd.DataFrame
-
-    :param macdonald_displacement_height:              Macdonald displacement height for each building.
-    :type macdonald_displacement_height:               pd.Series
-
-    :param vertical_distribution_of_building_heights:  Distribution of building heights for each building ata each height increment.
-    :type vertical_distribution_of_building_heights:   pd.DataFrame
-
-    :param building_geometry:                          Geometry field for the buildings.
-    :type building_geometry:                           pd.Series
-
-    :param target_crs:                                 Coordinate reference system field of the parent geometry.
-    :type target_crs:                                  CRS
-
-    :return:                                           Pandas DataFrame with all parameters merged together.
-    """
-
-    df = pd.concat([frontal_area_density, plan_area_density, rooftop_area_density], axis=1)
-
-    df[Settings.plan_area_fraction] = plan_area_fraction
-    df[Settings.mean_building_height] = mean_building_height
-    df[Settings.standard_deviation_of_building_heights] = standard_deviation_of_building_heights
-    df[Settings.area_weighted_mean_of_building_heights] = area_weighted_mean_of_building_heights
-    df[Settings.building_surface_area_to_plan_area_ratio] = building_surface_area_to_plan_area_ratio
-
-    df = pd.concat([df, frontal_area_index], axis=1)
-
-    df[Settings.complete_aspect_ratio] = complete_aspect_ratio
-    df[Settings.height_to_width_ratio] = height_to_width_ratio
-    df[Settings.sky_view_factor] = sky_view_factor
-    df[Settings.grimmond_oke_roughness_length] = grimmond_oke_roughness_length
-    df[Settings.grimmond_oke_displacement_height] = grimmond_oke_displacement_height
-
-    df[Settings.raupach_roughness_length_north] = raupach_roughness_length[
-        Settings.raupach_roughness_length_north
-    ]
-    df[Settings.raupach_displacement_height_north] = raupach_displacement_height[
-        Settings.raupach_displacement_height_north
-    ]
-    df[Settings.raupach_roughness_length_east] = raupach_roughness_length[
-        Settings.raupach_roughness_length_east
-    ]
-    df[Settings.raupach_displacement_height_east] = raupach_displacement_height[
-        Settings.raupach_displacement_height_east
-    ]
-    df[Settings.raupach_roughness_length_south] = raupach_roughness_length[
-        Settings.raupach_roughness_length_south
-    ]
-    df[Settings.raupach_displacement_height_south] = raupach_displacement_height[
-        Settings.raupach_displacement_height_south
-    ]
-    df[Settings.raupach_roughness_length_west] = raupach_roughness_length[
-        Settings.raupach_roughness_length_west
-    ]
-    df[Settings.raupach_displacement_height_west] = raupach_displacement_height[
-        Settings.raupach_displacement_height_west
-    ]
-
-    df = pd.concat([df, macdonald_roughness_length], axis=1)
-
-    df[Settings.macdonald_displacement_height] = macdonald_displacement_height
-
-    df = pd.concat([df, vertical_distribution_of_building_heights], axis=1)
-
-    df[Settings.geometry_field] = building_geometry
-
-    gdf = gpd.GeoDataFrame(df, geometry=Settings.geometry_field, crs=target_crs)
-
-    return gdf
-
-
-def numpy_to_binary(raster_to_numpy: np.ndarray) -> bytes:
-    """Turn the master numpy array containing all 132 aggregated parameters into a binary stream.
-
-    :param raster_to_numpy:         132 level numpy array with each level being an aggregated parameter.
-    :type raster_to_numpy:          np.ndarray
-
-    :return:                        Binary object containing the parameter data.
-    """
-
-    master_out = []
-
-    for i in range(len(raster_to_numpy)):
-        master_outi = bytes()
-        for j in range(len(raster_to_numpy[i])):
-            for k in range(len(raster_to_numpy[i][j])):
-                master_outi += struct.pack(">i", int(raster_to_numpy[i][j][k]))
-        master_out.append(master_outi)
-
-    master_out_final = bytes()
-
-    for i in range(len(master_out)):
-        master_out_final += master_out[i]
-
-    return master_out_final
-
-
-def orientation_to_neighbor(angle_in_degrees_to_neighbor: pd.Series) -> pd.Series:
-    """Determine the east-west or north-south orientation of the target building to its neighbors.
-
-    :param angle_in_degrees_to_neighbor:        Angle in degrees for each building interaction.
-    :type angle_in_degrees_to_neighbor:         pd.Series
-
-    :return:                                    Either the east-west or north-south orientation of the target building
-                                                to its neighbors.
-
-    """
-
-    return pd.Series(
-        np.where(
-            (
-                (
-                    (Settings.SOUTHEAST_DEGREES <= angle_in_degrees_to_neighbor)
-                    & (angle_in_degrees_to_neighbor <= Settings.DEGREES_IN_CIRCLE)
-                )
-                | (
-                    (Settings.START_OF_CIRCLE_DEGREES <= angle_in_degrees_to_neighbor)
-                    & (angle_in_degrees_to_neighbor <= Settings.NORTHEAST_DEGREES)
-                )
-            )
-            | (
-                (Settings.NORTHWEST_DEGREES <= angle_in_degrees_to_neighbor)
-                & (angle_in_degrees_to_neighbor <= Settings.SOUTHWEST_DEGREES)
-            ),
-            Settings.east_west,
-            Settings.north_south,
-        ),
-        index=angle_in_degrees_to_neighbor.index,
-    )
-
-
 def plan_area_density(
     building_plan_area: pd.Series, building_height: pd.Series, total_plan_area: pd.Series
 ) -> pd.DataFrame:
@@ -1367,55 +822,6 @@ def plan_area_fraction(building_plan_area: pd.Series, total_plan_area: pd.Series
     """
 
     return building_plan_area / total_plan_area
-
-
-def raster_to_numpy(aggregate_rasters: xr.Dataset) -> np.ndarray:
-    """Stack all 132 rasterized parameters into one numpy array for conversion to a binary file.
-
-    :param aggregate_rasters:                 Dataset with rasterized parameter values averaged at the defined resolution.
-    :type aggregate_rasters:                  xr.Dataset
-
-    :return:                                  132 level numpy array with each level being an aggregated parameter.
-    """
-
-    parameters = list(aggregate_rasters.keys())
-    parameters.remove("building_count")
-    rows = aggregate_rasters.dims["y"]
-    cols = aggregate_rasters.dims["x"]
-    master = np.zeros((132, rows, cols), dtype=np.float32)
-    i = 0
-    for parameter in parameters:
-        master[i] = aggregate_rasters[parameter].to_numpy()
-        i += 1
-
-    return master * 10**Settings.SCALING_FACTOR
-
-
-def rasterize_parameters(merge_parameters: gpd.GeoDataFrame) -> xr.Dataset:
-    """Rasterize parameters in preparation for conversion to numpy arrays. Raster will be of resolution Settings.DEFAULT_OUTPUT_RESOLUTION
-    and each cell will be the sum of each parameter value within. By default all_touched is True so that every building that is within a cell is
-    included in the sum.
-
-    :param merge_parameters:             Pandas.GeoDataFrame with all selected urban parameters for each building.
-    :type merge_parameters:              Pandas.GeoDataFrame
-
-    :return:                             Xr.Dataset containing rasterization of selected urban parameters.
-    """
-
-    merge_parameters["building_count"] = 1
-    measurements = merge_parameters.columns[merge_parameters.columns != Settings.geometry_field]
-    resolution = Settings.DEFAULT_OUTPUT_RESOLUTION
-    fill = Settings.DEFAULT_FILL_VALUE
-
-    return make_geocube(
-        vector_data=merge_parameters.rename(
-            columns={Settings.geometry_field: "geometry"}
-        ).set_geometry("geometry"),
-        measurements=measurements,
-        resolution=resolution,
-        fill=fill,
-        rasterize_function=partial(rasterize_image, all_touched=True, merge_alg=MergeAlg.add),
-    )
 
 
 def raupach_displacement_height(
@@ -1543,29 +949,6 @@ def sky_view_factor(
     """
 
     return np.cos(np.arctan(building_height / (0.5 * average_distance_between_buildings)))
-
-
-def standard_deviation_building_heights(
-    building_id: pd.Series, building_height_neighbor: pd.Series
-) -> pd.Series:
-    """Standard deviation of building heights for each target building when considering themselves
-    and those that are within their buffered area.
-
-    :param building_id:                         Building ID field.
-    :type building_id:                          pd.Series
-
-    :param building_height_neighbor:            Building height field for neighbors
-    :type building_height_neighbor:             pd.Series
-
-    :return:                                    Series of building heights for each target building in the buffered area
-
-    """
-
-    df = pd.DataFrame(
-        {Settings.id_field: building_id, Settings.neighbor_height_field: building_height_neighbor}
-    )
-
-    return df.groupby(Settings.id_field)[Settings.neighbor_height_field].std().fillna(0)
 
 
 def standard_deviation_of_building_heights(
@@ -1770,7 +1153,7 @@ def wall_angle_direction_length(building_geometry: pd.Series) -> pd.DataFrame:
 
 
 def wall_length(wall_angle_direction_length: pd.DataFrame) -> pd.DataFrame:
-    """Calculate the wall angle, direction, and length for each building in a GeoPandas GeoSeries.
+    """Calculate the wall length for each building in a GeoPandas GeoSeries.
 
     :param wall_angle_direction_length:                Wall angle, direction, and length for a series of buildings.
     :type wall_angle_direction_length:                 pd.DataFrame
@@ -1806,99 +1189,3 @@ def wall_length(wall_angle_direction_length: pd.DataFrame) -> pd.DataFrame:
         ],
         axis=1,
     )
-
-
-def write_binary(numpy_to_binary: bytes, raster_to_numpy: np.ndarray) -> None:
-    """Write the binary file that will be input to WRF.
-
-    :param numpy_to_binary:                 Binary object containing the parameter data.
-    :type numpy_to_binary:                  bytes
-
-    :param raster_to_numpy:                 132 level numpy array with each level being an aggregated parameter.
-    :type raster_to_numpy:                  np.ndarray
-    """
-
-    np.save("temporary.npy", numpy_to_binary)
-
-    rows = raster_to_numpy.shape[1]
-    cols = raster_to_numpy.shape[2]
-
-    first_y_index = "{:05d}".format(1)
-    second_y_index = "{:05d}".format(rows)
-
-    first_x_index = "{:05d}".format(1)
-    second_x_index = "{:05d}".format(cols)
-
-    out_binary_name = (
-        first_x_index + "-" + second_x_index + "." + first_y_index + "-" + second_y_index
-    )
-
-    with open("temporary.npy", "rb") as tile, open(out_binary_name, "wb") as tile2:
-        tile2.write(tile.read()[20 * 4 :])
-
-    os.remove("temporary.npy")
-
-
-def write_index(
-    raster_to_numpy: np.ndarray,
-    building_geometry: pd.Series,
-    target_crs: CRS,
-    index_filename: str = "index",
-) -> str:
-    """Write the index file that will accompany the output binary file.
-
-    :param raster_to_numpy:                 132 level numpy array with each level being an aggregated parameter.
-    :type raster_to_numpy:                  np.ndarray
-
-    :param building_geometry:               Geometry field for the buildings.
-    :type building_geometry:                pd.Series
-
-    :param target_crs:                      Coordinate reference system field of the parent geometry.
-    :type target_crs:                       crs
-    """
-
-    dy = float(Settings.DEFAULT_OUTPUT_RESOLUTION[0])
-    dx = float(Settings.DEFAULT_OUTPUT_RESOLUTION[1])
-
-    building_geometry = gpd.GeoSeries(building_geometry, crs=target_crs)
-
-    building_geometry_project = building_geometry.to_crs(crs=4326)
-
-    bounds = building_geometry_project.total_bounds
-
-    known_lat = bounds[1]
-    known_lon = bounds[0]
-
-    stdlon = (bounds[0] + bounds[2]) / 2
-
-    tile_x = raster_to_numpy.shape[2]
-    tile_y = raster_to_numpy.shape[1]
-
-    scale_factor = 10**-Settings.SCALING_FACTOR
-
-    with open(index_filename, "w") as index:
-        index.writelines(
-            [
-                "type=continuous\n",
-                "  projection=albers_nad83\n",
-                "  missing_value=-999900.\n",
-                "  dy=" + str(dy) + "\n",
-                "  dx=" + str(dx) + "\n",
-                "  known_x=1\n",
-                "  known_y=1\n",
-                "  known_lat=" + str(known_lat) + "\n",
-                "  known_lon=" + str(known_lon) + "\n",
-                "  truelat1=45.5\n",
-                "  truelat2=29.5\n",
-                "  stdlon=" + str(stdlon) + "\n",
-                "  wordsize=4\n",
-                "  endian=big\n",
-                "  signed=no\n",
-                "  tile_x=" + str(tile_x) + "\n",
-                "  tile_y=" + str(tile_y) + "\n",
-                "  tile_z=132\n",
-                '  units="dimensionless"\n',
-                "  scale_factor=" + str(scale_factor) + "\n",
-                '  description="Urban_Parameters"\n',
-            ]
-        )
