+++ tests/test_nodes.py
@@ -6,9 +6,8 @@ from dataclasses import dataclass
 import numpy as np
 import geopandas as gpd
 import pandas as pd
-from shapely.geometry import Point, Polygon, JOIN_STYLE
+from shapely.geometry import Polygon, JOIN_STYLE
 from typing import List
-import xarray as xr
 
 from naturf.driver import Model
 import naturf.nodes as nodes
@@ -21,112 +20,6 @@ class TestNodes(unittest.TestCase):
         "radius": 100,
         "cap_style": 1,
     }
-
-    def test_aggregate_rasters(self):
-        "Test that the function `aggregate_rasters()` returns the correct Xarray."
-        rasterize_parameters = xr.Dataset(
-            data_vars=dict(
-                var0=(["x", "y"], np.array([[0, 1], [2, 3]])),
-                var1=(["x", "y"], np.array([[10, 15.5], [314159, 75]])),
-                building_count=(["x", "y"], np.array([[0, 1], [10, 75]])),
-            ),
-            coords=dict(
-                lat=(["x"], [0, 1]),
-                lon=(["y"], [0, 1]),
-            ),
-        )
-
-        expected = xr.Dataset(
-            data_vars=dict(
-                var0=(["x", "y"], np.array([[0.0, 1.0], [0.2, 0.04]])),
-                var1=(["x", "y"], np.array([[math.inf, 15.5], [31415.9, 1.0]])),
-                building_count=(["x", "y"], np.array([[0.0, 1.0], [1.0, 1.0]])),
-            ),
-            coords=dict(
-                lat=(["x"], [0, 1]),
-                lon=(["y"], [0, 1]),
-            ),
-        )
-        actual = nodes.aggregate_rasters(rasterize_parameters)
-        xr.testing.assert_equal(expected, actual)
-
-    def test_angle_in_degrees_to_neighbor(self):
-        "Test that the function `angle_in_degrees_to_neighbor()` returns the correct angle."
-
-        @dataclass
-        class TestCase:
-            name: str
-            target_input: List[Point]
-            neighbor_input: List[Point]
-            expected: List[int | float | str]
-
-        testcases = [
-            TestCase(
-                name="same_centroid",
-                target_input=[Point(1, 1), Point(0, 0)],
-                neighbor_input=[Point(1, 1), Point(0, 0)],
-                expected=[0.0, 0.0],
-            ),
-            TestCase(
-                name="east",
-                target_input=[Point(0, 0)],
-                neighbor_input=[Point(1, 0)],
-                expected=[0.0],
-            ),
-            TestCase(
-                name="west",
-                target_input=[Point(0, 0)],
-                neighbor_input=[Point(-1, 0)],
-                expected=[180.0],
-            ),
-            TestCase(
-                name="north",
-                target_input=[Point(0, 0)],
-                neighbor_input=[Point(0, 1)],
-                expected=[90.0],
-            ),
-            TestCase(
-                name="south",
-                target_input=[Point(0, 0)],
-                neighbor_input=[Point(0, -1)],
-                expected=[270.0],
-            ),
-            TestCase(
-                name="northeast",
-                target_input=[Point(0, 0)],
-                neighbor_input=[Point(10, 10 * np.sqrt(3))],
-                expected=[60.0],
-            ),
-            TestCase(
-                name="northwest",
-                target_input=[Point(0, 0)],
-                neighbor_input=[Point(-10, 10 * np.sqrt(3))],
-                expected=[120.0],
-            ),
-            TestCase(
-                name="southeast",
-                target_input=[Point(0, 0)],
-                neighbor_input=[Point(10, -10 * np.sqrt(3))],
-                expected=[300.0],
-            ),
-            TestCase(
-                name="southwest",
-                target_input=[Point(0, 0)],
-                neighbor_input=[Point(-10, -10 * np.sqrt(3))],
-                expected=[240.0],
-            ),
-        ]
-
-        for case in testcases:
-            actual = nodes.angle_in_degrees_to_neighbor(
-                gpd.GeoSeries(case.target_input), gpd.GeoSeries(case.neighbor_input)
-            )
-            expected = pd.Series(case.expected)
-            pd.testing.assert_series_equal(
-                expected,
-                actual,
-                f"angle_in_degrees_to_neighbor test {case.name} failed, expected {expected}, actual {actual}",
-            )
 
     def test_area_weighted_mean_of_building_heights(self):
         "Test that the function `area_weighted_mean_of_building_heights()` returns the correct value."
@@ -983,173 +876,6 @@ class TestNodes(unittest.TestCase):
             f"mean_building_height test failed, expected {expected}, actual {actual}",
         )
 
-    def test_merge_parameters(self):
-        """Test the function `merge_parameters()` to ensure it outputs the right type and shape GeoDataFrame."""
-
-        frontal_area_density = pd.DataFrame(
-            {
-                Settings.frontal_area_north: [0.1, 0.2],
-                Settings.frontal_area_east: [0.3, 0.4],
-                Settings.frontal_area_south: [0.5, 0.6],
-                Settings.frontal_area_west: [0.7, 0.8],
-            }
-        )
-        plan_area_density = pd.DataFrame({Settings.plan_area_density: [0.5, 0.6]})
-        rooftop_area_density = pd.DataFrame({Settings.rooftop_area_density: [0.9, 1.0]})
-        plan_area_fraction = pd.Series([0.5, 0.5])
-        mean_building_height = pd.Series([10, 20])
-        standard_deviation_of_building_heights = pd.Series([2, 3])
-        area_weighted_mean_of_building_heights = pd.Series([12, 22])
-        building_surface_area_to_plan_area_ratio = pd.Series([1.2, 1.3])
-        frontal_area_index = pd.DataFrame(
-            {
-                Settings.frontal_area_index_north: [0.0, 0.1],
-                Settings.frontal_area_index_east: [0.2, 0.3],
-                Settings.frontal_area_index_south: [0.4, 0.5],
-                Settings.frontal_area_index_west: [0.6, 0.7],
-            }
-        )
-        complete_aspect_ratio = pd.Series([1.5, 1.6])
-        height_to_width_ratio = pd.Series([1.7, 1.8])
-        sky_view_factor = pd.Series([0.9, 0.8])
-        grimmond_oke_roughness_length = pd.Series([0.7, 0.6])
-        grimmond_oke_displacement_height = pd.Series([0.5, 0.4])
-        raupach_roughness_length = pd.DataFrame(
-            {
-                Settings.raupach_roughness_length_north: [0.3, 0.2],
-                Settings.raupach_roughness_length_east: [0.2, 0.3],
-                Settings.raupach_roughness_length_south: [0.4, 0.5],
-                Settings.raupach_roughness_length_west: [0.6, 0.7],
-            }
-        )
-        raupach_displacement_height = pd.DataFrame(
-            {
-                Settings.raupach_displacement_height_north: [0.0, 0.1],
-                Settings.raupach_displacement_height_east: [0.2, 0.3],
-                Settings.raupach_displacement_height_south: [0.4, 0.5],
-                Settings.raupach_displacement_height_west: [0.6, 0.7],
-            }
-        )
-        macdonald_roughness_length = pd.DataFrame(
-            {
-                Settings.macdonald_roughness_length_north: [0.0, 0.1],
-                Settings.macdonald_roughness_length_east: [0.2, 0.3],
-                Settings.macdonald_roughness_length_south: [0.4, 0.5],
-                Settings.macdonald_roughness_length_west: [0.6, 0.7],
-            }
-        )
-        macdonald_displacement_height = pd.Series([0.8, 0.9])
-        vertical_distribution_of_building_heights = pd.DataFrame(
-            {Settings.vertical_distribution_of_building_heights: [10, 20]}
-        )
-        building_geometry = pd.Series([Point(0, 0), Point(1, 1)])
-        target_crs = "epsg:3857"
-
-        result = nodes.merge_parameters(
-            frontal_area_density,
-            plan_area_density,
-            rooftop_area_density,
-            plan_area_fraction,
-            mean_building_height,
-            standard_deviation_of_building_heights,
-            area_weighted_mean_of_building_heights,
-            building_surface_area_to_plan_area_ratio,
-            frontal_area_index,
-            complete_aspect_ratio,
-            height_to_width_ratio,
-            sky_view_factor,
-            grimmond_oke_roughness_length,
-            grimmond_oke_displacement_height,
-            raupach_roughness_length,
-            raupach_displacement_height,
-            macdonald_roughness_length,
-            macdonald_displacement_height,
-            vertical_distribution_of_building_heights,
-            building_geometry,
-            target_crs,
-        )
-
-        assert isinstance(result, gpd.GeoDataFrame), "Output is not a GeoDataFrame"
-        assert list(result.columns) == [
-            "frontal_area_north",
-            "frontal_area_east",
-            "frontal_area_south",
-            "frontal_area_west",
-            "plan_area_density",
-            "rooftop_area_density",
-            "plan_area_fraction",
-            "mean_building_height",
-            "standard_deviation_of_building_heights",
-            "area_weighted_mean_of_building_heights",
-            "building_surface_area_to_plan_area_ratio",
-            "frontal_area_index_north",
-            "frontal_area_index_east",
-            "frontal_area_index_south",
-            "frontal_area_index_west",
-            "complete_aspect_ratio",
-            "height_to_width_ratio",
-            "sky_view_factor",
-            "grimmond_oke_roughness_length",
-            "grimmond_oke_displacement_height",
-            "raupach_roughness_length_north",
-            "raupach_displacement_height_north",
-            "raupach_roughness_length_east",
-            "raupach_displacement_height_east",
-            "raupach_roughness_length_south",
-            "raupach_displacement_height_south",
-            "raupach_roughness_length_west",
-            "raupach_displacement_height_west",
-            "macdonald_roughness_length_north",
-            "macdonald_roughness_length_east",
-            "macdonald_roughness_length_south",
-            "macdonald_roughness_length_west",
-            "macdonald_displacement_height",
-            "vertical_distribution_of_building_heights",
-            "building_geometry",
-        ], "Output columns are not as expected"
-        assert result.crs == target_crs, "Output CRS is not as expected"
-
-    def test_numpy_to_binary(self):
-        """Test the function `numpy_to_binary()` to ensure it outputs the right type and length binary file."""
-
-        test_array = np.array([[[1, 2, 3], [4, 5, 6]], [[7, 8, 9], [10, 11, 12]]])
-
-        binary_output = nodes.numpy_to_binary(test_array)
-
-        assert isinstance(binary_output, bytes), "Output is not of type 'bytes'"
-        assert len(binary_output) == 48, "Binary output length is not as expected"
-
-    def test_orientation_to_neighbor(self):
-        """Test that the function `orientation_to_neighbor()` returns either `north_south` or `east_west` correctly."""
-
-        @dataclass
-        class TestCase:
-            name: str
-            input: List[int | float]
-            expected: List[int | str]
-
-        east_west = Settings.east_west
-        north_south = Settings.north_south
-
-        testcases = [
-            TestCase(name="zero_degrees", input=[0.0, -0.0], expected=[east_west, east_west]),
-            TestCase(name="north_south", input=[90, 270], expected=[north_south, north_south]),
-            TestCase(
-                name="east_west",
-                input=[45, 135, 225, 315, 360],
-                expected=[east_west, east_west, east_west, east_west, east_west],
-            ),
-        ]
-
-        for case in testcases:
-            actual = nodes.orientation_to_neighbor(pd.Series(case.input))
-            expected = pd.Series(case.expected)
-            pd.testing.assert_series_equal(
-                expected,
-                actual,
-                f"orientation_to_neighbor test {case.name} failed, expected {expected}, actual {actual}",
-            )
-
     def test_plan_area_density(self):
         """Test that the function `plan_area_density()` returns the correct value."""
 
@@ -1202,52 +928,6 @@ class TestNodes(unittest.TestCase):
             actual,
             f"plan_area_fraction test failed, expected {expected}, actual {actual}",
         )
-
-    def test_raster_to_numpy(self):
-        """Test the function `raster_to_numpy()` to ensure it outputs the right type and shape numpy array."""
-
-        aggregate_rasters = xr.Dataset(
-            {
-                "parameter1": (("x", "y"), np.random.rand(5, 5)),
-                "parameter2": (("x", "y"), np.random.rand(5, 5)),
-                "building_count": (("x", "y"), np.random.randint(0, 10, (5, 5))),
-            },
-            coords={"x": range(5), "y": range(5)},
-        )
-
-        result = nodes.raster_to_numpy(aggregate_rasters)
-
-        assert isinstance(result, np.ndarray), "Output is not a numpy array"
-        assert result.shape == (132, 5, 5), "Output shape is not as expected"
-        assert result.dtype == np.float32, "Output dtype is not np.float32"
-
-    def test_rasterize_parameters(self):
-        """Test the function `rasterize_parameters()` to ensure it outputs the right type and shape xr.Dataset."""
-
-        merge_parameters = gpd.GeoDataFrame(
-            {
-                "parameter1": [1, 2, 3],
-                "parameter2": [4, 5, 6],
-                Settings.geometry_field: [Point(0, 0), Point(1, 1), Point(2, 2)],
-            }
-        )
-
-        result = nodes.rasterize_parameters(merge_parameters)
-
-        assert isinstance(result, xr.Dataset), "Output is not an xr.Dataset"
-        assert list(result.keys()) == [
-            "parameter1",
-            "parameter2",
-            "building_count",
-        ], "Output keys are not as expected"
-        assert result["parameter1"].shape == (
-            1,
-            1,
-        ), "Output shape for 'parameter1' is not as expected"
-        assert result["parameter2"].shape == (
-            1,
-            1,
-        ), "Output shape for 'parameter2' is not as expected"
 
     def test_raupach_displacement_height(self):
         """Test that the function `raupach_displacement_height()` returns the correct value for each cardinal direction."""
@@ -1683,71 +1363,6 @@ class TestNodes(unittest.TestCase):
                 actual,
                 f"wall_length test {case.name} failed, expected {expected}, actual {actual}",
             )
-
-    def test_write_binary(self):
-        """Test that the function `write_binary()` writes a binary file correctly."""
-
-        raster_to_numpy = np.zeros((132, 10, 10))
-        test_binary_filename = "00001-00010.00001-00010"
-        numpy_to_binary = raster_to_numpy.tobytes()
-
-        nodes.write_binary(numpy_to_binary, raster_to_numpy)
-
-        assert os.path.exists(test_binary_filename), "Binary file was not created."
-        with open(test_binary_filename, "rb") as binary_file:
-            content = binary_file.read()
-            assert len(content) == 105648, "Content length is not as expected"
-
-        os.remove(test_binary_filename)
-
-    def test_write_index(self):
-        """Test that the function `write_index()` writes an index file and contains the correct values."""
-
-        raster_to_numpy = np.zeros((132, 10, 10))
-        polygon1 = Polygon([[0, 0], [0, 1], [1, 1], [1, 0]])
-        polygon2 = Polygon([[3, 3], [3, 4], [4, 4], [4, 3]])
-        building_geometry = pd.Series([polygon1, polygon2])
-        target_crs = "epsg:3857"
-        test_index_filename = "test_index"
-
-        nodes.write_index(
-            raster_to_numpy, building_geometry, target_crs, index_filename=test_index_filename
-        )
-
-        assert os.path.exists(test_index_filename), "Index file was not created."
-        with open("index", "r") as index:
-            content = index.read()
-            assert "type=continuous" in content, "Index file type is not as expected."
-            assert "projection=albers_nad83" in content, "Index file projection is not as expected."
-            assert (
-                "missing_value=-999900." in content
-            ), "Index file missing_value is not as expected."
-            assert "dy=100." in content, "Index file dy is not as expected."
-            assert "dx=100." in content, "Index file dx is not as expected."
-            assert "known_x=1" in content, "Index file known_x is not as expected."
-            assert "known_y=1" in content, "Index file known_y is not as expected."
-            assert (
-                "known_lat=38.860881738862304" in content
-            ), "Index file known_lat is not as expected."
-            assert (
-                "known_lon=-77.04903814481347" in content
-            ), "Index file known_lon is not as expected."
-            assert "truelat1=45.5" in content, "Index file truelat1 is not as expected."
-            assert "truelat2=29.5" in content, "Index file truelat2 is not as expected."
-            assert "stdlon=-77.03451915155296" in content, "Index file stdlon is not as expected."
-            assert "wordsize=4" in content, "Index file wordsize is not as expected."
-            assert "endian=big" in content, "Index file endian is not as expected."
-            assert "signed=no" in content, "Index file signed is not as expected."
-            assert "tile_x=26" in content, "Index file tile_x is not as expected."
-            assert "tile_y=25" in content, "Index file tile_y is not as expected."
-            assert "tile_z=132" in content, "Index file tile_z is not as expected."
-            assert 'units="dimensionless"' in content, "Index file units is not as expected."
-            assert "scale_factor=0.0001" in content, "Index file scale_factor is not as expected."
-            assert (
-                'description="Urban_Parameters"' in content
-            ), "Index file description is not as expected."
-
-        os.remove(test_index_filename)
 
 
 if __name__ == "__main__":
